feat: use generate query patterns
